requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
pydub==0.25.1
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
requests==2.32.3
SpeechRecognition==3.10.4
typing_extensions==4.12.2
urllib3==2.2.3
